HW1_All_Codes/q4_code_and_results/requirements.txt
numpy>=1.26
pillow>=10.0
notebook>=7.0
ipykernel>=6.0
